pyproject.toml
[project]
name = "village"
version = "0.1.0"
description = "Keeping an agent under control"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "google-genai>=1.29.0",
]
